requirement.txt
asgiref==3.2.7
Django==3.0.5
django-tinymce==2.8.0
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
